requirements.txt
python-telegram-bot==21.10
ollama==0.4.7
python-dotenv==1.1.1
httpx==0.28.1
google-generativeai==0.8.4